feat(audit-logs): extend feature to other resources

+++ apps/server/src/bl/users/user.bl.ts
@@ -1,9 +1,10 @@
 import { UserRepository } from '../../dal/userRepository';
 import bcrypt from 'bcrypt';
-import { Role, User } from '@OpsiMate/shared';
+import { AuditActionType, AuditResourceType, Role, User } from '@OpsiMate/shared';
+import { AuditBL } from '../audit/audit.bl';
 
 export class UserBL {
-    constructor(private userRepo: UserRepository) {}
+    constructor(private userRepo: UserRepository, private auditBL: AuditBL) {}
 
     async register(email: string, fullName: string, password: string): Promise<User> {
         const userCount = await this.userRepo.countUsers();
@@ -14,19 +15,52 @@ export class UserBL {
         const result = await this.userRepo.createUser(email, hash, fullName, 'admin');
         const user = await this.userRepo.getUserById(result.lastID);
         if (!user) throw new Error('User creation failed');
+
+				await this.auditBL.logAction({
+					// registered by self
+					actionType: AuditActionType.CREATE,
+					resourceType: AuditResourceType.USER,
+					resourceId: user.id.toString(),
+					userId: user.id,
+					resourceName: user.fullName,
+					userName: user.fullName,
+					details: JSON.stringify({ role: user.role, email: user.email }),
+				});
         return user;
     }
 
-    async createUser(email: string, fullName: string, password: string, role: Role): Promise<User> {
+    async createUser(email: string, fullName: string, password: string, role: Role, creator: { creatorId: number; creatorName: string }): Promise<User> {
         const hash = await bcrypt.hash(password, 10);
         const result = await this.userRepo.createUser(email, hash, fullName, role);
         const user = await this.userRepo.getUserById(result.lastID);
         if (!user) throw new Error('User creation failed');
+
+				await this.auditBL.logAction({
+					actionType: AuditActionType.CREATE,
+					resourceType: AuditResourceType.USER,
+					resourceId: user.id.toString(),
+					userId: creator.creatorId,
+					resourceName: user.fullName,
+					userName: creator.creatorName,
+					details: JSON.stringify({ role: user.role, email: user.email }),
+				});
         return user;
     }
 
-    async updateUserRole(email: string, newRole: Role): Promise<void> {
-        await this.userRepo.updateUserRole(email, newRole);
+    async updateUserRole(email: string, newRole: Role, updater: { updaterId: number; updaterName: string }): Promise<void> {
+			await this.userRepo.updateUserRole(email, newRole);
+
+			const user = await this.userRepo.getUserByEmail(email);
+			if (!user) throw new Error('User not found');
+			await this.auditBL.logAction({
+				actionType: AuditActionType.UPDATE,
+				resourceType: AuditResourceType.USER,
+				resourceId: user.id.toString(),
+				userId: updater.updaterId,
+				resourceName: user.fullName,
+				userName: updater.updaterName,
+				details: JSON.stringify({ 'New Role': user.role, email: user.email }),
+			});
     }
 
     async login(email: string, password: string): Promise<User> {
@@ -59,8 +93,20 @@ export class UserBL {
         return await this.userRepo.getAllUsers()
     }
 
-    async deleteUser(id: number): Promise<void> {
-        await this.userRepo.deleteUser(id);
+    async deleteUser(id: number, deleter: { deleterId: number; deleterName: string }): Promise<void> {
+			const user = await this.userRepo.getUserById(id);
+			if (!user) throw new Error('User not found');
+	
+			await this.userRepo.deleteUser(id);
+			await this.auditBL.logAction({
+				actionType: AuditActionType.DELETE,
+				resourceType: AuditResourceType.USER,
+				resourceId: user.id.toString(),
+				userId: deleter.deleterId,
+				resourceName: user.fullName,
+				userName: deleter.deleterName,
+				details: JSON.stringify({ role: user.role, email: user.email }),
+			});
     }
 
     async getUserById(id: number): Promise<User | null> {
@@ -81,11 +127,20 @@ export class UserBL {
             passwordHash = await bcrypt.hash(newPassword, 10);
         }
         
-        await this.userRepo.updateUserProfile(id, fullName, passwordHash);
-        const updatedUser = await this.userRepo.getUserById(id);
-        if (!updatedUser) {
-            throw new Error('User not found');
-        }
-        return updatedUser;
+				await this.userRepo.updateUserProfile(id, fullName, passwordHash);
+				const updatedUser = await this.userRepo.getUserById(id);
+				if (!updatedUser) throw new Error('User not found');
+
+				await this.auditBL.logAction({
+					// updated by self
+					actionType: AuditActionType.UPDATE,
+					resourceType: AuditResourceType.USER,
+					resourceId: updatedUser.id.toString(),
+					userId: updatedUser.id,
+					resourceName: updatedUser.fullName,
+					userName: updatedUser.fullName,
+					details: JSON.stringify({ fullName: updatedUser.fullName }),
+				});
+				return updatedUser;
     }
 }
